n84 rocket.io

+++ app/src/main/java/com/viatom/checkme/bean/PedBean.kt
@@ -6,9 +6,9 @@ data class PedBean(
     var date: Date = Date(),
     var timeString: String = "",
     var step: Int = 0,
-    var dis: Int = 0,
-    var speed: Int = 0,
-    var cal: Int = 0,
-    var fat: Int = 0,
+    var dis: Float = 0f,
+    var speed: Float = 0f,
+    var cal: Float = 0f,
+    var fat: Float = 0f,
     var time: Int = 0
 )
